Added player and ball prefab, refactored field generator.
