+++ TeamBuildingGame/Assets/Scenes/Main.unity
@@ -130,9 +130,9 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 89541187}
-  - component: {fileID: 89541186}
   - component: {fileID: 89541188}
   - component: {fileID: 89541189}
+  - component: {fileID: 89541186}
   m_Layer: 0
   m_Name: PlayingField
   m_TagString: Untagged
@@ -149,13 +149,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 89541185}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 208b47a15fbb3cb47aa2fb8fb9bc1fb2, type: 3}
+  m_Script: {fileID: 11500000, guid: c44252b2043f9f5419b3faf7cef6d7d9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   p1: {x: 0, y: 0, z: 0}
-  p2: {x: 0, y: 400, z: 0}
-  p3: {x: 400, y: 0, z: 0}
-  p4: {x: 400, y: 400, z: 0}
+  p2: {x: 0, y: 5, z: 0}
+  p3: {x: 5, y: 0, z: 0}
+  p4: {x: 5, y: 5, z: 0}
 --- !u!4 &89541187
 Transform:
   m_ObjectHideFlags: 0
@@ -300,3 +300,141 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1001 &2151501399476000630
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 1.46
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 1.07
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -1.2757258
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 2
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145335, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2151501400236145355, guid: 59fa0ce6ef71257468f3649d25d7e713,
+        type: 3}
+      propertyPath: m_Name
+      value: Ball
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 59fa0ce6ef71257468f3649d25d7e713, type: 3}
+--- !u!1001 &6173906396138195347
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 6173906396210492117, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_Name
+      value: Player
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 1.54
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 2.57
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 3
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6173906396210492121, guid: 920963b816785504194569c4fade5156,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 920963b816785504194569c4fade5156, type: 3}
